// packages/adapter-travis/src/types.ts
export type TravisSpec = Job & {
  notifications?: {
    webhooks?:
      | string[]
      | {
        urls?: string | string[];
        [k: string]: any;
      };
    slack?:
      | SlackRoom
      | {
        rooms?: SlackRoom[];
        on_pull_requests?: boolean;
        template?: string[];
        on_success?: NotificationFrequency;
        on_failure?: NotificationFrequency;
      };
    email?:
      | string[]
      | false
      | {
        recipients?: string[];
        on_success?: 'always' | 'never' | 'change';
        on_failure?: 'always' | 'never' | 'change';
        [k: string]: any;
      };
    irc?:
      | string
      | string[]
      | {
        channels?: string[];
        channel_key?: string;
        nick?: string;
        password?: string;
        template?: string[];
        on_success?: 'always' | 'never' | 'change';
        on_failure?: 'always' | 'never' | 'change';
        skip_join?: boolean;
        use_notice?: boolean;
        [k: string]: any;
      };
  };
  matrix?: {
    exclude?: Job[];
    include?: Job[];
    allow_failures?: Job[];
    /**
     * If some rows in the build matrix are allowed to fail, the build won’t be marked as finished until they have completed. To mark the build as finished as soon as possible, add fast_finish: true
     */
    fast_finish?: boolean;
  };
  jobs?: {
    include?: (Job & {
      /**
       * The name of the build stage
       */
      stage?: string;
      [k: string]: any;
    })[];
    allow_failures?: Job[];
  };
  /**
   * Specifies the order of build stages
   */
  stages?: (
    | string
    | {
      name?: string;
        /**
         * Specifies a condition for the stage
         */
      if?: string;
    })[];
  [k: string]: any;
}
// ...
export type SlackRoom = string
export type NotificationFrequency = 'always' | 'never' | 'change'

export interface Job {
  language?:
    | 'android'
    | 'bash'
    | 'c'
    | 'c++'
    | 'clojure'
    | 'cpp'
    | 'crystal'
    | 'csharp'
    | 'd'
    | 'dart'
    | 'dartlang'
    | 'elixir'
    | 'elm'
    | 'erlang'
    | 'generic'
    | 'go'
    | 'golang'
    | 'groovy'
    | 'haskell'
    | 'haxe'
    | 'java'
    | 'javascript'
    | 'julia'
    | 'jvm'
    | 'minimal'
    | 'nix'
    | 'node'
    | 'node.js'
    | 'node_js'
    | 'nodejs'
    | 'obj-c'
    | 'obj_c'
    | 'objective-c'
    | 'objective_c'
    | 'perl'
    | 'perl6'
    | 'php'
    | 'python'
    | 'r'
    | 'ruby'
    | 'rust'
    | 'scala'
    | 'sh'
    | 'shell'
    | 'smalltalk'
// ...
}